+++ apps/minifront/src/abort-loader.ts
@@ -1,4 +1,9 @@
-import { penumbra, throwIfPraxNotConnected, throwIfPraxNotInstalled } from './prax';
+import { penumbra } from './penumbra';
+import {
+  PenumbraClient,
+  PenumbraNotInstalledError,
+  PenumbraProviderNotConnectedError,
+} from '@penumbra-zone/client';
 
 /**
  * Retry test, resolving `true`, or resolving `false` if timeout reached.
@@ -23,15 +28,27 @@ const retry = async (fn: () => boolean, ms = 500, rate = Math.max(ms / 10, 50)) 
     }, ms);
   });
 
+const throwIfProviderNotConnected = () => {
+  if (!penumbra.connected) {
+    throw new PenumbraProviderNotConnectedError();
+  }
+};
+
+const throwIfProviderNotInstalled = () => {
+  if (!Object.keys(PenumbraClient.getProviders()).length) {
+    throw new PenumbraNotInstalledError();
+  }
+};
+
 /**
  * Resolves fast if Prax is connected, or rejects if Prax is not connected after
  * timeout. This is a temporary solution until loaders properly await Prax
  * connection.
  */
 export const abortLoader = async (): Promise<null> => {
-  await throwIfPraxNotInstalled();
+  throwIfProviderNotInstalled();
   await retry(() => Boolean(penumbra.connected));
-  throwIfPraxNotConnected();
+  throwIfProviderNotConnected();
 
   // Loaders are required to return a value, even if it's null. By returning
   // `null` here, we can use this loader directly in the router.