+++ packages/wasm/crate/src/keys.rs
@@ -1,7 +1,5 @@
-use std::str::FromStr;
-
 use anyhow;
-use penumbra_keys::keys::{Bip44Path, SeedPhrase, SpendKey};
+use penumbra_keys::keys::{AddressIndex, Bip44Path, SeedPhrase, SpendKey};
 use penumbra_keys::{Address, FullViewingKey};
 use penumbra_proof_params::{
     CONVERT_PROOF_PROVING_KEY, DELEGATOR_VOTE_PROOF_PROVING_KEY, OUTPUT_PROOF_PROVING_KEY,
@@ -10,6 +8,7 @@ use penumbra_proof_params::{
 use penumbra_proto::core::keys::v1 as pb;
 use penumbra_proto::DomainType;
 use rand_core::OsRng;
+use std::str::FromStr;
 use wasm_bindgen::prelude::*;
 
 use crate::error::WasmResult;
@@ -135,4 +134,50 @@ pub fn is_controlled_address(full_viewing_key: &[u8], address: &[u8]) -> WasmRes
 
 pub fn is_controlled_inner(fvk: &FullViewingKey, address: &Address) -> bool {
     fvk.address_index(address).is_some()
+}
+
+#[wasm_bindgen(getter_with_clone)]
+pub struct ForwardingAddrResponse {
+    /// A noble address that will be used for registration on the noble network
+    pub noble_addr_bech32: String,
+    /// Byte representation of the noble forwarding address. Used for broadcasting cosmos message.
+    pub noble_addr_bytes: Vec<u8>,
+    /// The penumbra address that a deposit to the noble address with forward to
+    /// Vec encoded `pb::Address`
+    pub penumbra_addr_bytes: Vec<u8>,
+}
+
+/// Generates an address that can be used as a forwarding address for Noble
+/// Returns: Uint8Array representing encoded Address
+#[wasm_bindgen]
+pub fn get_noble_forwarding_addr(
+    sequence: u16,
+    full_viewing_key: &[u8],
+    channel: &str,
+    account: Option<u32>,
+) -> WasmResult<ForwardingAddrResponse> {
+    let fvk: FullViewingKey = FullViewingKey::decode(full_viewing_key)?;
+    let penumbra_addr = forwarding_addr_inner(sequence, account, &fvk);
+    let noble_addr = penumbra_addr.noble_forwarding_address(channel);
+    Ok(ForwardingAddrResponse {
+        noble_addr_bech32: noble_addr.to_string(),
+        noble_addr_bytes: noble_addr.bytes(),
+        penumbra_addr_bytes: penumbra_addr.encode_to_vec(),
+    })
+}
+
+/// Noble Randomizer: [0xff; 10] followed by LE16(sequence)
+pub fn forwarding_addr_inner(sequence: u16, account: Option<u32>, fvk: &FullViewingKey) -> Address {
+    let mut randomizer: [u8; 12] = [0xff; 12]; // Initialize all 12 bytes to 0xff
+    let seq_bytes = sequence.to_le_bytes();
+    randomizer[10..].copy_from_slice(&seq_bytes); // Replace the last 2 bytes with seq_bytes
+
+    let index = AddressIndex {
+        account: account.unwrap_or_default(),
+        randomizer,
+    };
+
+    let (address, _dtk) = fvk.incoming().payment_address(index);
+
+    address
 }