+++ packages/wasm/src/asset.ts
@@ -1,12 +1,13 @@
 import { AssetId } from '@penumbra-zone/protobuf/penumbra/core/asset/v1/asset_pb';
-import { get_asset_id_inner } from '../wasm/index.js';
+import { get_asset_id } from '../wasm/index.js';
 
 /**
  * Converts a base denom name string to an `AssetId` with inner binary field
  * @param altBaseDenom an asset's base denomination name
- * @returns the appropriate `AssetId` with binary inner field
+ * @returns the appropriate `AssetId`
  */
 export const assetIdFromBaseDenom = (altBaseDenom: string) => {
-  const inner = get_asset_id_inner(new AssetId({ altBaseDenom }).toBinary());
-  return new AssetId({ inner });
+  const inputBytes = new AssetId({ altBaseDenom }).toBinary();
+  const outputBytes = get_asset_id(inputBytes);
+  return AssetId.fromBinary(outputBytes);
 };