+++ apps/minifront/CHANGELOG.md
@@ -1,5 +1,40 @@
 # minifront
 
+## 6.23.1
+
+### Patch Changes
+
+- @penumbra-zone/perspective@34.0.0
+- @penumbra-zone/ui@11.0.1
+
+## 6.23.0
+
+### Minor Changes
+
+- 45c6751: Add multiwallet support to minifront
+
+### Patch Changes
+
+- Updated dependencies [9d68f48]
+- Updated dependencies [deb04f5]
+- Updated dependencies [e543db4]
+- Updated dependencies [dad8165]
+  - @penumbra-zone/ui@11.0.0
+  - @penumbra-zone/protobuf@6.2.0
+  - @penumbra-zone/perspective@33.0.0
+  - @penumbra-zone/bech32m@9.0.0
+  - @penumbra-zone/client@20.0.0
+  - @penumbra-zone/getters@19.0.0
+  - @penumbra-zone/types@25.0.0
+  - @penumbra-zone/crypto-web@26.0.0
+
+## 6.22.4
+
+### Patch Changes
+
+- Updated dependencies [4885420]
+  - @penumbra-zone/ui@10.0.2
+
 ## 6.22.3
 
 ### Patch Changes